[msc3787_test.go] Reconfigure this whole file of tests to use a single deployment.
This only works because of(previous work, plus) the fact that new Blueprint with 3 homeservers uses a layout compatible with the BlueprintAlice, BlueprintFederationTwoLocalOneRemote, BlueprintOneToOneRoom and BlueprintFederationOneToOneRoom. Comments were added to specific subtests to detail which users were used. Since the original TestRestrictedRoomsRemoteJoinFailOver used a custom 3 homeserver blueprint, this was the most expensive deployment in this file, so it's the one that's to be used.