[restricted_room_hierarchy_test.go] Refactor tests.
1. Wrap all tests in t.Run().
2. Share deployment among all tests.
3. Parallel tests.

Original Test times: 89 seconds(including 62 seconds for blueprints and 15 seconds for containers), net of 11 seconds.
Sequential Test times: 61 seconds(including 40 seconds for blueprints and 8 seconds for containers), net of 13 seconds.
Parallel Test times: 57 seconds(including 39 seconds for blueprints and 8 seconds for containers), net of 10 seconds.

Conclusion: A smell placebo. Clearly less container spin up is a win, and parallel will help with keeping the system busy, but was this worth it? Each of these tests only take 2-3 seconds to actually complete, but now we are doing both at once. I think a synapse workers run would be the only place this would be helpful.